% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotGridMap.R
\name{plotGridMap}
\alias{plotGridMap}
\title{Plots a filled grid map of US states, represented as hexagons}
\usage{
plotGridMap(x, state_var = "state.abb", fill_var, alpha_var = NULL,
  label_var = NULL, label_color = "black")
}
\arguments{
\item{x}{data.frame with data to plot}

\item{state_var}{name of of column in \code{x} that contains state abbreviations (matching \code{base::state.abb()})}

\item{fill_var}{column from \code{x} to use for color fill; can be numeric, factor, or character (treated as factor by ggplot)}

\item{alpha_var}{column from \code{x} to use for fill alpha (NULL by default, for no alpha); can be numeric, factor, or character (treated as factor by ggplot)}

\item{label_var}{name of of column in \code{x} to use for labels; if NULL, then no labels}

\item{label_color}{color for label text}
}
\value{
Returns a ggplot
}
\description{
Plots a filled grid map of US states, represented as hexagons
}
\examples{
# a data.frame of states with random categorical value
my_dat <- data.frame(state.abb, value=sample(LETTERS[1:5], 50, replace=T))

# build grid map plot
my_dat <- data.frame(state.abb = c(state.abb, 'DC'),
                     value=sample(LETTERS[1:5], 51, replace=T))

# note you can manipulate the resultant object as you would any ggplot object
my_grid_map +
 scale_fill_brewer(type='qual') +
 labs(title = 'States by Category', fill = 'Category')
}

